add specular lighting

+++ src/material_system.h
@@ -10,7 +10,7 @@ struct MaterialDescr {
     u64     diffuseMap = INVALID_TEXTURE;
     u64     specularMap = INVALID_TEXTURE;
     u64     normalMap = INVALID_TEXTURE;
-    u32     padding = 0;
+    u64     padding = 0;
 };
 
 static_assert(sizeof(MaterialDescr) % 16 == 0, "MaterialDescr should be padded to 16 bytes");
